src/requirements.txt
atom==0.4.1
automationhat==0.1.0
blinker==1.3
blinkt==0.1.1
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.4.16
chardet==3.0.4
click==6.6
colorama==0.3.7
cryptography==1.7.1
Cython==0.28.2
Django==1.11
django-cors-headers==2.2.0
djangorestframework==3.6.2
drumhat==0.1.0
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
future==0.16.0
gpiozero==1.4.1
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
lxkeymap==0.1
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.2.1
mote==0.0.3
motephat==0.0.2
mysqlclient==1.3.7
numpy==1.14.2
oauthlib==2.0.1
pantilthat==0.0.5
phatbeat==0.1.0
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
pyasn1==0.1.9
pycrypto==2.6.1
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
PyMySQL==0.8.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
pyserial==3.2.1
pytz==2018.4
pyxdg==0.25
rainbowhat==0.0.2
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.6.3
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.0
sense-hat==2.2.0
simplejson==3.10.0
six==1.10.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
touchphat==0.0.1
twython==3.4.0
unicornhathd==0.0.3
urllib3==1.19.1
virtualenv==15.2.0
Werkzeug==0.11.15
